Only swallow `require` errors about missing modules
We don’t care about these errors, because all it means is that the stylesheet is not defined. We can just return the default empty object in those cases.

Other errors, however, we want to surface. This came up when `emotion` was throwing an error about needing access to the document.