chore(deps): upgrade to `emotion` 10
BREAKING CHANGE: While neither the public API of `emotion` nor the addon changed, the Babel plugin has slightly less options that it did before, which could break some people's usage of the Ember addon. Because of this, it makes more sense to cut a breaking change. Note that, if you're not configuing the Babel plugin yourself, this is likely an upgrade you can perform withotu changing your code at all.